Finally fixed lat and lon showing as doubles instead of integers

+++ libs/backends/geolocation/geolocation.h
@@ -25,8 +25,8 @@ typedef enum {
 typedef struct location_t {
     int id;
     char* name;
-    int longitude;
-    int latitude;
+    double longitude;
+    double latitude;
     double elevation;
     char* feature_code;
     char* country_code;
